arch/x86_64: map kernel image with proper permissions
Previously the kernel was mapped as a 128MiB identity map of address 0,
now each PHDR is mapped on its own and is mapped with the proper flags.

This also enables CR0.WP, so writes to read-only pages cause a page
fault in the kernel.

+++ kernel/arch/x86_64/mm/vmm.h
@@ -7,14 +7,14 @@
 #define VMM_FLAG_MASK (0xFFF | (1ull << 63))
 #define VMM_ADDR_MASK ~(VMM_FLAG_MASK)
 
-#define VMM_FLAG_PRESENT	(1<<0)
-#define VMM_FLAG_WRITE		(1<<1)
-#define VMM_FLAG_USER		(1<<2)
-#define VMM_FLAG_PAT0		(1<<3)
-#define VMM_FLAG_PAT1		(1<<4)
-#define VMM_FLAG_PAT2		(1<<7)
-#define VMM_FLAG_DIRTY		(1<<5)
-#define VMM_FLAG_LARGE		(1<<7)
+#define VMM_FLAG_PRESENT	(1ull<<0)
+#define VMM_FLAG_WRITE		(1ull<<1)
+#define VMM_FLAG_USER		(1ull<<2)
+#define VMM_FLAG_PAT0		(1ull<<3)
+#define VMM_FLAG_PAT1		(1ull<<4)
+#define VMM_FLAG_PAT2		(1ull<<7)
+#define VMM_FLAG_DIRTY		(1ull<<5)
+#define VMM_FLAG_LARGE		(1ull<<7)
 #define VMM_FLAG_NX		(1ull<<63)
 
 typedef struct {
@@ -33,9 +33,9 @@ void *vmm_offs_to_virt(pt_off_t);
 
 void vmm_init(void);
 
-int vmm_map_pages(pt_t *, void *, void *, size_t, int);
+int vmm_map_pages(pt_t *, void *, void *, size_t, unsigned long);
 int vmm_unmap_pages(pt_t *, void *, size_t);
-int vmm_update_perms(pt_t *, void *, size_t, int);
+int vmm_update_perms(pt_t *, void *, size_t, unsigned long);
 
 uintptr_t vmm_get_entry(pt_t *, void *);
